i2c: xiic: Use i2c_add_numbered_adapter
In 'xiic: Read the bus number from the device tree' we added support for
reading a bus-id field from the device tree. However, i2c_add_adapter
ignores this value and dynamically assigns the bus number. Use
i2c_add_numbered_adapter to resolve this issue. Provide a default value
of -1 to get the same behavior as i2c_add_adapter if a bus-id isn't
specified in the device tree.

Signed-off-by: Jeff Westfahl <jeff.westfahl@ni.com>