macb: Use the PHY via PHY handle
If available, we'll attempt to locate the PHY via the PHY handle,
falling back to the previous discovery method if a PHY handle is not
available.

By using the OF PHY handle, we can specify for a MAC to use a PHY that
is anywhere else in device tree, including under another MAC. This
allows us to share PHYs between MACs.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Acked-by: Ben Shelton <ben.shelton@ni.com>
Acked-by: Gratian Crisan <gratian.crisan@ni.com>
Natinst-ReviewBoard-ID: 84559