HACK net/macb: Totally disable TX checksum offloading
Since hardware checksumming is not supported, reset the TXCOEN bit in
DMACFG. If unitialized, this bit causes checksum offloading to be left
on in hardware, with software thinking it is off, and for a checksum
offloading bug in hardware to cause small UDP packets (<= 2 bytes) to be
sent with an incorrect checksum (0xffff).

We can't fix this problem by zeroing the checksum field in outgoing
UDPv4 packets as implemented by Jeff Westfahl or Xilinx in their xemacps
driver because it breaks packet forwarding (where it is illegal to
modify the checksum of forwarded packets with unknown validity). We
don't yet know a way to fix this problem other than to not use hardware
checksum offloading.

This commit will need to be reworked before it is upstreamable. We just
need a fix in the LabVIEW RT Linux kernel right now for release.
Ideally, we'd continue to work with upstream to solve the problem, but
this is an emergency and we need to test and ship right away.

Signed-off-by: Mihai Neagu <mihai.neagu@ni.com>
Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Reviewed-by: Josh Cartwright <joshc@ni.com>
Natinst-CAR-ID: 516373
Natinst-ReviewBoard-ID: 97988
(joshc: also hack up hw_features to never advertise hw csum support)
Signed-off-by: Josh Cartwright <joshc@ni.com>