net: macb: Free IRQ on close
When fpgaperipheral brings the network down, it previously freed the
device IRQ with free_irq(). This free was removed because it caused a
multiple IRQ free when macb_remove() was called. Since fpgaperipheral no
longer freed the IRQ, when the FPGA up notification was received,
macb_open would try to use the interrupt it already had, which failed.
This lead to the network device not coming back up (and remaining
closed).

This patch solves the problem by freeing the interrupt in close using
devm_free_irq() instead of free_irq(). This prevents the double free
from happening that 2820f05d7b87 originally tried to fix.

The reason this is a valid solution, even though upstream doesn't free
the IRQ during close, is because we actually still need to free the IRQ
in close. For some of our boards, an IRQ line (used for macb) between
the CPU and FPGA floats when the FPGA is not programmed. As such, it
makes sense to free the IRQ (which also disables it) while the FPGA is
down.

The following cases were tested.

Avoid Double Free IRQ after device_release_driver()
---------------------------------------------------
1. We macb_remove the interface or use device_release_driver().
2. A double free of the IRQ does not happen.

Avoid Re-requesting IRQs (CAR 524447)
-------------------------------------
1. We receive an FPGA_PERIPHERAL_DOWN notification. This calls
   macb_close().
2. We receive an FPGA_PERIPHERAL_UP notification. This calls
   macb_open().
3. We do not request an IRQ we already have ownership of.

Signed-off-by: Jaeden Amero <jaeden.amero@ni.com>
Fixes: ea0adb9e47a8 ("net: macb: Add support for device polling")
Reviewed-by: Josh Cartwright <joshc@ni.com>
Reviewed-by: Ben Shelton <ben.shelton@ni.com>
Natinst-CAR-ID: 524447
Natinst-ReviewBoard-ID: 95951