clk: si5351: allow manual configuration of pll frequencies
Even when setting all clock outputs to be a pll-master, sometimes we
end up with a less than optimal choice for the PLL frequency. The case
in question is:

            |  desired | actual (pll frequency: 897024000)
   clkout0: | 12800000 | 12814628
   clkout1: | 80000000 | 80091428
   clkout2: | 13107200 | 13122179
   clkout4: | 10240000 | 10251702
   clkout7: | 12288000 | 12288000

This patch allows manual specification of PLL frequencies; in the above
case, it allows us to specify a frequency of 737280000 which lets us
hit all five frequencies exactly.

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Acked-by: James Minor <james.minor@ni.com>
Natinst-ReviewBoard-ID: 191413