Revert "net: dsa: mv88e6xxx: steer parent clock to adjust PHC frequency"
This reverts commit 4e8b9c904fadc343f92a7b61ccb78ad40b3b230e.

This commit added functionality that was unused in our implementation,
and was never upstreamed. The upstream PTP subsystem doesn't really have
the notion of adjusting an external oscillator (see comments in [1]),
so in order to to more easily integrate the timecounter/cyclecounter
pieces from upstream, lets just get rid of this.

[1]: https://lore.kernel.org/lkml/20170929094323.cwh2ubv4odknlyot@localhost/

Signed-off-by: Brandon Streiff <brandon.streiff@ni.com>
Acked-by: Erik Hons <erik.hons@ni.com>
Acked-by: Kyle Roeschley <kyle.roeschley@ni.com>
Natinst-ReviewBoard-ID: 311588