ni-bluefinlc: Add device tree
Device tree for Bluefin LC devices are for Low-cost
Ethernet cDAQ-9183 and cDAQ-9187.
The changes here is to update the model name, NIDEVCODE
and removal of dsa switch port 2 (sw1) since Bluefin LC
devices only have a single ethernet port. Otherwise, the
rest of the device tree remains the same as Bluefin.

Signed-off-by: Kuhanh Murugasen Krishnan <kuhanh.murugasen.krishnan@ni.com>